test(rest-api-client): refactor structures of tests for AppClient and RecordClient
